Create README.md
The objective of this project is to streamline the lunch ordering process for employees by offering an up-to-date online menu, 
enabling pre-orders to reduce waiting times, and facilitating desk deliveries. 
The system will enhance employee satisfaction, minimize food wastage, and optimize operational efficiency within the canteen. 
We identified the impacted stakeholders,drew up stakeholders analysis matrix, identified the main problems within the current process, developed a detailed business case to 
convince the Project Sponsor, developed RACI Matrix for the implementation of the project, highlighted the main features
to be developed in the system, created the AS-IN and TO-BE process Maps, translated the TO-BE requirements into user stories and created a Jira Board and Uploaded the user stories